+++ saleor/graphql/checkout/mutations.py
@@ -275,6 +275,14 @@ def group_quantity_by_variants(lines: List[Dict[str, Any]]) -> List[int]:
     return list(variant_quantity_map.values())
 
 
+def validate_checkout_email(checkout: models.Checkout):
+    if not checkout.email:
+        raise ValidationError(
+            "Checkout email must be set.",
+            code=CheckoutErrorCode.EMAIL_NOT_SET.value,
+        )
+
+
 class CheckoutLineInput(graphene.InputObjectType):
     quantity = graphene.Int(required=True, description="The number of items purchased.")
     variant_id = graphene.ID(required=True, description="ID of the product variant.")
@@ -1143,12 +1151,26 @@ class CheckoutEmailUpdate(BaseMutation):
         error_type_class = CheckoutError
         error_type_field = "checkout_errors"
 
+    @staticmethod
+    def clean_email(email):
+        if not email:
+            raise ValidationError(
+                {
+                    "email": ValidationError(
+                        "This field cannot be blank.",
+                        code=CheckoutErrorCode.REQUIRED.value,
+                    )
+                }
+            )
+
     @classmethod
     def perform_mutation(cls, _root, info, email, checkout_id=None, token=None):
         # DEPRECATED
         validate_one_of_args_is_in_mutation(
             CheckoutErrorCode, "checkout_id", checkout_id, "token", token
         )
+
+        cls.clean_email(email)
 
         if token:
             checkout = get_checkout_by_token(token)
@@ -1648,6 +1670,8 @@ class CheckoutComplete(BaseMutation):
                     )
                 raise e
 
+            validate_checkout_email(checkout)
+
             manager = info.context.plugins
             lines = fetch_checkout_lines(checkout)
             validate_variants_in_checkout_lines(lines)
@@ -1721,6 +1745,8 @@ class CheckoutAddPromoCode(BaseMutation):
             checkout = cls.get_node_or_error(
                 info, checkout_id or token, only_type=Checkout, field="checkout_id"
             )
+
+        validate_checkout_email(checkout)
 
         manager = info.context.plugins
         discounts = info.context.discounts