+++ saleor/graphql/checkout/tests/test_checkout_promo_codes.py
@@ -845,6 +845,19 @@ def test_checkout_add_promo_code_invalidate_shipping_method(
     assert shipping_method_id not in data["checkout"]["availableShippingMethods"]
 
 
+def test_checkout_add_promo_code_no_checkout_email(
+    api_client, checkout_with_item, voucher
+):
+    checkout_with_item.email = None
+    checkout_with_item.save(update_fields=["email"])
+
+    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    data = _mutate_checkout_add_promo_code(api_client, variables)
+
+    assert data["errors"]
+    assert data["errors"][0]["code"] == CheckoutErrorCode.EMAIL_NOT_SET.name
+
+
 MUTATION_CHECKOUT_REMOVE_PROMO_CODE = """
     mutation($token: UUID, $promoCode: String, $promoCodeId: ID) {
         checkoutRemovePromoCode(