+++ saleor/order/tests/test_order_actions_create_fulfillments_for_returned_products.py
@@ -6,14 +6,9 @@ from prices import Money, TaxedMoney
 from ...plugins.manager import get_plugins_manager
 from ...tests.utils import flush_post_commit_hooks
 from ...warehouse.models import Allocation, Stock
-from .. import (
-    FulfillmentLineData,
-    FulfillmentStatus,
-    OrderEvents,
-    OrderLineData,
-    OrderOrigin,
-)
+from .. import FulfillmentLineData, FulfillmentStatus, OrderEvents, OrderOrigin
 from ..actions import create_fulfillments_for_returned_products
+from ..fetch import OrderLineInfo
 from ..models import Fulfillment, FulfillmentLine
 
 
@@ -45,7 +40,7 @@ def test_create_return_fulfillment_only_order_lines(
         order=order_with_lines,
         payment=payment,
         order_lines=[
-            OrderLineData(line=line, quantity=2, replace=False)
+            OrderLineInfo(line=line, quantity=2, replace=False)
             for line in order_lines_to_return
         ],
         fulfillment_lines=[],
@@ -119,7 +114,7 @@ def test_create_return_fulfillment_only_order_lines_with_refund(
         order=order_with_lines,
         payment=payment,
         order_lines=[
-            OrderLineData(line=line, quantity=2, replace=False)
+            OrderLineInfo(line=line, quantity=2, replace=False)
             for line in order_lines_to_return
         ],
         fulfillment_lines=[],
@@ -191,7 +186,7 @@ def test_create_return_fulfillment_only_order_lines_included_shipping_costs(
         order=order_with_lines,
         payment=payment,
         order_lines=[
-            OrderLineData(line=line, quantity=2, replace=False)
+            OrderLineInfo(line=line, quantity=2, replace=False)
             for line in order_lines_to_return
         ],
         fulfillment_lines=[],
@@ -263,7 +258,7 @@ def test_create_return_fulfillment_only_order_lines_with_replace_request(
     lines_count = order_with_lines.lines.count()
     quantity_to_replace = 2
     order_lines_data = [
-        OrderLineData(line=line, quantity=2, replace=False)
+        OrderLineInfo(line=line, quantity=2, replace=False)
         for line in order_lines_to_return
     ]