+++ saleor/order/tests/test_order_utils.py
@@ -1,3 +1,4 @@
+import copy
 from decimal import Decimal
 from unittest.mock import Mock
 
@@ -9,8 +10,9 @@ from ...giftcard import GiftCardEvents
 from ...giftcard.models import GiftCardEvent
 from ...order.interface import OrderTaxedPricesData
 from ...plugins.manager import get_plugins_manager
-from .. import OrderLineData, OrderStatus
+from .. import OrderStatus
 from ..events import OrderEvents
+from ..fetch import OrderLineInfo
 from ..models import Order, OrderEvent
 from ..utils import (
     add_gift_cards_to_order,
@@ -18,6 +20,7 @@ from ..utils import (
     change_order_line_quantity,
     get_valid_shipping_methods_for_order,
     match_orders_with_new_user,
+    update_taxes_for_order_line,
     update_taxes_for_order_lines,
 )
 
@@ -46,7 +49,7 @@ def test_change_quantity_generates_proper_event(
 
     line = order_with_lines.lines.last()
     line.quantity = previous_quantity
-    line_info = OrderLineData(
+    line_info = OrderLineInfo(
         line=line,
         quantity=line.quantity,
         variant=line.variant,
@@ -94,7 +97,7 @@ def test_change_quantity_update_line_fields(
 ):
     # given
     line = order_with_lines.lines.last()
-    line_info = OrderLineData(
+    line_info = OrderLineInfo(
         line=line,
         quantity=line.quantity,
         variant=line.variant,
@@ -446,3 +449,41 @@ def test_add_gift_cards_to_order_no_checkout_user(
         },
         "order_id": order.id,
     }
+
+
+def test_update_taxes_for_order_line_deleted_variant(order_with_lines):
+    # given
+    order_line = order_with_lines.lines.first()
+    order_line_unchanged_copy = copy.deepcopy(order_line)
+    unit_discount_amount = Decimal("2.00")
+
+    unit_price = TaxedMoney(net=Money("10.23", "USD"), gross=Money("15.80", "USD"))
+    total_price = TaxedMoney(net=Money("30.34", "USD"), gross=Money("36.49", "USD"))
+    tax_rate = Decimal("0.23")
+    discount = TaxedMoney(
+        net=Money(unit_discount_amount, "USD"), gross=Money(unit_discount_amount, "USD")
+    )
+    unit_price_data = OrderTaxedPricesData(
+        undiscounted_price=unit_price + discount,
+        price_with_discounts=unit_price,
+    )
+    total_price_data = OrderTaxedPricesData(
+        undiscounted_price=total_price + discount,
+        price_with_discounts=total_price,
+    )
+    manager = Mock(
+        calculate_order_line_unit=Mock(return_value=unit_price_data),
+        calculate_order_line_total=Mock(return_value=total_price_data),
+        get_order_line_tax_rate=Mock(return_value=tax_rate),
+    )
+
+    # when
+    order_line.variant = None
+    update_taxes_for_order_line(order_line, order_with_lines, manager, True)
+
+    # then
+    assert order_line.unit_price == order_line_unchanged_copy.unit_price
+    assert order_line.total_price == order_line_unchanged_copy.total_price
+    assert order_line.tax_rate == order_line_unchanged_copy.tax_rate
+    assert order_line.undiscounted_unit_price == order_line_unchanged_copy.unit_price
+    assert order_line.undiscounted_total_price == order_line_unchanged_copy.total_price