+++ saleor/warehouse/tests/test_stock_management.py
@@ -5,7 +5,7 @@ from django.db.models import Sum
 from django.db.models.functions import Coalesce
 
 from ...core.exceptions import InsufficientStock
-from ...order import OrderLineData
+from ...order.fetch import OrderLineInfo
 from ...order.models import OrderLine
 from ...plugins.manager import get_plugins_manager
 from ...tests.utils import flush_post_commit_hooks
@@ -28,7 +28,7 @@ def test_allocate_stocks(order_line, stock, channel_USD):
     stock.quantity = 100
     stock.save(update_fields=["quantity"])
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=50)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=50)
 
     allocate_stocks(
         [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
@@ -58,10 +58,10 @@ def test_allocate_stocks_multiple_lines(order_line, order, product, stock, chann
 
     quantity_1 = 50
     quantity_2 = 5
-    line_data_1 = OrderLineData(
+    line_data_1 = OrderLineInfo(
         line=order_line, variant=order_line.variant, quantity=quantity_1
     )
-    line_data_2 = OrderLineData(
+    line_data_2 = OrderLineInfo(
         line=order_line_2, variant=variant_2, quantity=quantity_2
     )
 
@@ -86,7 +86,7 @@ def test_allocate_stock_many_stocks(order_line, variant_with_many_stocks, channe
     variant = variant_with_many_stocks
     stocks = variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=5)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=5)
     allocate_stocks(
         [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
     )
@@ -105,7 +105,7 @@ def test_allocate_stock_with_reservations(
     variant = variant_with_many_stocks
     stocks = variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=3)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=3)
     allocate_stocks(
         [line_data],
         COUNTRY_CODE,
@@ -128,7 +128,7 @@ def test_allocate_stock_insufficient_stock_due_to_reservations(
     variant = variant_with_many_stocks
     variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=5)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=5)
 
     with pytest.raises(InsufficientStock):
         allocate_stocks(
@@ -152,7 +152,7 @@ def test_allocate_stock_many_stocks_partially_allocated(
     variant = allocated_line.variant
     stocks = variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=3)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=3)
     allocate_stocks(
         [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
     )
@@ -169,7 +169,7 @@ def test_allocate_stock_partially_allocated_insufficient_stocks(
     variant = allocated_line.variant
     stocks = variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=6)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=6)
     with pytest.raises(InsufficientStock):
         allocate_stocks(
             [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
@@ -186,7 +186,7 @@ def test_allocate_stocks_no_channel_shipping_zones(order_line, stock, channel_US
     stock.quantity = 100
     stock.save(update_fields=["quantity"])
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=50)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=50)
     with pytest.raises(InsufficientStock):
         allocate_stocks(
             [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
@@ -199,7 +199,7 @@ def test_allocate_stock_insufficient_stocks(
     variant = variant_with_many_stocks
     stocks = variant.stocks.all()
 
-    line_data = OrderLineData(line=order_line, variant=order_line.variant, quantity=10)
+    line_data = OrderLineInfo(line=order_line, variant=order_line.variant, quantity=10)
     with pytest.raises(InsufficientStock):
         allocate_stocks(
             [line_data], COUNTRY_CODE, channel_USD.slug, manager=get_plugins_manager()
@@ -228,10 +228,10 @@ def test_allocate_stock_insufficient_stocks_for_multiple_lines(
 
     quantity_1 = 100
     quantity_2 = 100
-    line_data_1 = OrderLineData(
+    line_data_1 = OrderLineInfo(
         line=order_line, variant=order_line.variant, quantity=quantity_1
     )
-    line_data_2 = OrderLineData(
+    line_data_2 = OrderLineInfo(
         line=order_line_2, variant=variant_2, quantity=quantity_2
     )
 
@@ -259,7 +259,7 @@ def test_deallocate_stock(allocation):
 
     deallocate_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line, quantity=80, variant=stock.product_variant
             )
         ],
@@ -281,7 +281,7 @@ def test_deallocate_stock_when_quantity_less_than_zero(allocation):
 
     deallocate_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line, quantity=80, variant=stock.product_variant
             )
         ],
@@ -303,7 +303,7 @@ def test_deallocate_stock_partially(allocation):
 
     deallocate_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line, quantity=50, variant=stock.product_variant
             )
         ],
@@ -322,7 +322,7 @@ def test_deallocate_stock_many_allocations(
     order_line = order_line_with_allocation_in_many_stocks
 
     deallocate_stock(
-        [OrderLineData(line=order_line, quantity=3, variant=order_line.variant)],
+        [OrderLineInfo(line=order_line, quantity=3, variant=order_line.variant)],
         manager=get_plugins_manager(),
     )
 
@@ -337,7 +337,7 @@ def test_deallocate_stock_many_allocations_partially(
     order_line = order_line_with_allocation_in_many_stocks
 
     deallocate_stock(
-        [OrderLineData(line=order_line, quantity=1, variant=order_line.variant)],
+        [OrderLineInfo(line=order_line, quantity=1, variant=order_line.variant)],
         manager=get_plugins_manager(),
     )
 
@@ -392,7 +392,7 @@ def test_increase_stock_with_new_allocation(order_line, stock):
 @pytest.mark.parametrize("quantity", (19, 20))
 def test_increase_allocations(quantity, allocation):
     order_line = allocation.order_line
-    order_line_info = OrderLineData(
+    order_line_info = OrderLineInfo(
         line=order_line,
         quantity=quantity,
         variant=order_line.variant,
@@ -421,7 +421,7 @@ def test_increase_allocations(quantity, allocation):
 
 def test_increase_allocation_insufficient_stock(allocation):
     order_line = allocation.order_line
-    order_line_info = OrderLineData(
+    order_line_info = OrderLineInfo(
         line=order_line,
         quantity=21,
         variant=order_line.variant,
@@ -477,7 +477,7 @@ def test_decrease_stock(allocation):
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line,
                 quantity=50,
                 variant=stock.product_variant,
@@ -504,7 +504,7 @@ def test_decrease_stock_without_stock_update(quantity, expected_allocated, alloc
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line,
                 quantity=quantity,
                 variant=stock.product_variant,
@@ -538,13 +538,13 @@ def test_decrease_stock_multiple_lines(allocations):
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation_1.order_line,
                 quantity=50,
                 variant=allocation_1.order_line.variant,
                 warehouse_pk=warehouse_pk_1,
             ),
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation_2.order_line,
                 quantity=20,
                 variant=allocation_2.order_line.variant,
@@ -570,7 +570,7 @@ def test_decrease_stock_partially(allocation):
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line,
                 quantity=80,
                 variant=stock.product_variant,
@@ -595,7 +595,7 @@ def test_decrease_stock_many_allocations(
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=order_line,
                 quantity=3,
                 variant=order_line.variant,
@@ -620,7 +620,7 @@ def test_decrease_stock_many_allocations_partially(
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=order_line,
                 quantity=2,
                 variant=order_line.variant,
@@ -649,7 +649,7 @@ def test_decrease_stock_more_then_allocated(
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=order_line,
                 quantity=4,
                 variant=order_line.variant,
@@ -677,7 +677,7 @@ def test_decrease_stock_insufficient_stock(allocation):
     with pytest.raises(InsufficientStock):
         decrease_stock(
             [
-                OrderLineData(
+                OrderLineInfo(
                     line=allocation.order_line,
                     quantity=50,
                     variant=stock.product_variant,
@@ -753,7 +753,7 @@ def test_decrease_stock_with_out_of_stock_webhook_triggered(
 
     decrease_stock(
         [
-            OrderLineData(
+            OrderLineInfo(
                 line=allocation.order_line,
                 quantity=50,
                 variant=stock.product_variant,
@@ -776,7 +776,7 @@ def test_allocate_preorders(
     channel_listing.preorder_quantity_threshold = 100
     channel_listing.save(update_fields=["preorder_quantity_threshold"])
 
-    line_data = OrderLineData(line=order_line, variant=variant, quantity=50)
+    line_data = OrderLineInfo(line=order_line, variant=variant, quantity=50)
 
     allocate_preorders([line_data], channel_USD.slug)
 
@@ -801,7 +801,7 @@ def test_allocate_preorders_with_allocation(
     channel_listing.save(update_fields=["preorder_quantity_threshold"])
 
     quantity_to_allocate = 2
-    line_data = OrderLineData(
+    line_data = OrderLineInfo(
         line=order_line, variant=variant, quantity=quantity_to_allocate
     )
 
@@ -823,7 +823,7 @@ def test_allocate_preorders_insufficient_stocks_channel_threshold(
     channel_listing = variant.channel_listings.get(channel_id=channel_USD.id)
     channel_listings = variant.channel_listings.all()
 
-    line_data = OrderLineData(
+    line_data = OrderLineInfo(
         line=order_line,
         variant=variant,
         quantity=channel_listing.preorder_quantity_threshold + 1,
@@ -851,7 +851,7 @@ def test_allocate_preorders_insufficient_stocks_global_threshold(
     )
     available_preorder_quantity = variant.preorder_global_threshold - global_allocation
 
-    line_data = OrderLineData(
+    line_data = OrderLineInfo(
         line=order_line,
         variant=variant,
         quantity=available_preorder_quantity + 1,
@@ -873,7 +873,7 @@ def test_allocate_preorders_with_channel_reservations(
     channel_listing.preorder_quantity_threshold = 5
     channel_listing.save(update_fields=["preorder_quantity_threshold"])
 
-    line_data = OrderLineData(line=order_line, variant=variant, quantity=5)
+    line_data = OrderLineInfo(line=order_line, variant=variant, quantity=5)
 
     with pytest.raises(InsufficientStock):
         allocate_preorders(
@@ -899,7 +899,7 @@ def test_allocate_preorders_with_global_reservations(
     variant.preorder_global_threshold = 5
     variant.save()
 
-    line_data = OrderLineData(line=order_line, variant=variant, quantity=5)
+    line_data = OrderLineInfo(line=order_line, variant=variant, quantity=5)
 
     with pytest.raises(InsufficientStock):
         allocate_preorders(